refactor(cli): share diff fetching and marker upsert helpers
Extract getGitHubDiff and the marker-based comment upsert from the
classify command into packages/cli/src/github.ts so the upcoming ai-review
command can reuse them. Adds AI_REVIEW_COMMENT_MARKER +
upsertAiReviewComment alongside the existing classify marker; classify
behavior is unchanged.
